fix timout saving mrs for project 11
The issue where the calls time out for large projects is back. Saving the commits inside the mergeRequests seems to be causing the issue because it is a large amount of data. This call is now seperated the call that sets the project. The new db desing should also fix this isseu