Fix bugs in projections
The projection needs to be done before aggregation. I think this is because the @JsonProperty annotation saves the data as the annotation name and once we added this for more variables the names in the aggregations didnt match the projection lets it know wht fields I am referring to and solves the issue.